+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 <!-- and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). -->
 
+## [Unreleased]
+### Fixed
+- Restore helpful error message when `browser_launch_timeout` is too small (lost during browser pool introduction)
+
 ## [0.11.1] 2026-02-10
 ### Fixed
 - Fix regression: Config validation error if `browser_pools` not set. Commit [987828e]